requirements.txt
bidict==0.21.2
click==8.0.1
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.1
Flask-SocketIO==5.1.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
Pillow==8.3.1
python-engineio==4.2.1
python-socketio==5.4.0
six==1.16.0
Werkzeug==2.0.1
